fix: Consolidate metadata and viewport configuration into layout.tsx for improved organization
